currently all tests pass in 32bit and 64bit. Still needs work on the hash.

+++ src/test/test_base.h
@@ -4,6 +4,6 @@
 #include "robin_hood.h"
 #include "sfc64.h"
 
-using FlatMap = robin_hood::flat_map<int, int>;
-using NodeMap = robin_hood::node_map<int, int>;
+using FlatMap = robin_hood::flat_map<uint64_t, uint64_t>;
+using NodeMap = robin_hood::node_map<uint64_t, uint64_t>;
 using Rng = sfc64;
